+++ NSeleneTests/Examples/SharedDriver/WithPageObjects/TodoMvc_Should.cs
@@ -13,8 +13,8 @@ namespace NSelene.Tests.Examples.SharedDriver.WithPageObjects
         public void initDriver()
         {
             new DriverManager().SetUpDriver(
-                // new ChromeConfig(), version: "Latest"
-                new ChromeConfig(), version: "89.0.4389.23"
+                new ChromeConfig(), version: "Latest"
+                // new ChromeConfig(), version: "89.0.4389.23"
             );
             var options = new ChromeOptions();
             options.AddArguments("headless");