+++ NSeleneTests/Integration/SharedDriver/Configuration_Specs.cs
@@ -17,8 +17,8 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneElementSpec
         public void InitConfiguration()
         {            
             new DriverManager().SetUpDriver(
-                // new ChromeConfig(), version: "Latest"
-                new ChromeConfig(), version: "89.0.4389.23"
+                new ChromeConfig(), version: "Latest"
+                // new ChromeConfig(), version: "89.0.4389.23"
             );
 
             var options = new ChromeOptions();