+++ NSeleneTests/Integration/SharedDriver/SeleneElement_SendKeys_Specs.cs
@@ -39,7 +39,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
             Assert.AreEqual(
                 "before and after", 
                 Configuration.Driver
-                .FindElement(By.TagName("input")).GetProperty("value")
+                .FindElement(By.TagName("input")).GetDomProperty("value")
             );
             Assert.Greater(afterCall, beforeCall.AddSeconds(0.3));
             Assert.Less(afterCall, beforeCall.AddSeconds(0.7));
@@ -102,7 +102,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
             StringAssert.Contains(
                 "empty.html", 
                 Configuration.Driver
-                .FindElement(By.CssSelector("[type=file]")).GetProperty("value")
+                .FindElement(By.CssSelector("[type=file]")).GetDomProperty("value")
             );
         }
 
@@ -135,7 +135,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
             Assert.AreEqual(
                 "before and after", 
                 Configuration.Driver
-                .FindElement(By.TagName("input")).GetProperty("value")
+                .FindElement(By.TagName("input")).GetDomProperty("value")
             );
             Assert.Greater(afterCall, beforeCall.AddSeconds(0.3));
             Assert.Less(afterCall, beforeCall.AddSeconds(0.7));
@@ -176,7 +176,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
                 Assert.AreEqual(
                     "before ", 
                     Configuration.Driver
-                    .FindElement(By.TagName("input")).GetProperty("value")
+                    .FindElement(By.TagName("input")).GetDomProperty("value")
                 );
             }
         }
@@ -230,7 +230,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
             Assert.AreEqual(
                 "before and after", 
                 Configuration.Driver
-                .FindElement(By.TagName("input")).GetProperty("value")
+                .FindElement(By.TagName("input")).GetDomProperty("value")
             );
         }