+++ NSeleneTests/Integration/SharedDriver/SeleneElement_SetValue_Specs.cs
@@ -39,7 +39,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
             Assert.AreEqual(
                 "overwritten", 
                 Configuration.Driver
-                .FindElement(By.TagName("input")).GetProperty("value")
+                .FindElement(By.TagName("input")).GetDomProperty("value")
             );
             Assert.Greater(afterCall, beforeCall.AddSeconds(0.3));
             Assert.Less(afterCall, beforeCall.AddSeconds(0.6));
@@ -147,7 +147,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
                 Assert.AreEqual(
                     "", 
                     Configuration.Driver
-                    .FindElement(By.TagName("input")).GetProperty("value")
+                    .FindElement(By.TagName("input")).GetDomProperty("value")
                 );
             }
         }
@@ -192,7 +192,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
                 Assert.AreEqual(
                     "", 
                     Configuration.Driver
-                    .FindElement(By.TagName("input")).GetProperty("value")
+                    .FindElement(By.TagName("input")).GetDomProperty("value")
                 );
             }
         }
@@ -226,7 +226,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
             Assert.AreEqual(
                 "overwritten", 
                 Configuration.Driver
-                .FindElement(By.TagName("input")).GetProperty("value")
+                .FindElement(By.TagName("input")).GetDomProperty("value")
             );
             Assert.Greater(afterCall, beforeCall.AddSeconds(0.3));
             Assert.Less(afterCall, beforeCall.AddSeconds(0.6));
@@ -267,7 +267,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
                 Assert.AreEqual(
                     "initial", 
                     Configuration.Driver
-                    .FindElement(By.TagName("input")).GetProperty("value")
+                    .FindElement(By.TagName("input")).GetDomProperty("value")
                 );
             }
         }
@@ -307,7 +307,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
                 Assert.AreEqual(
                     "initial", 
                     Configuration.Driver
-                    .FindElement(By.TagName("input")).GetProperty("value")
+                    .FindElement(By.TagName("input")).GetDomProperty("value")
                 );
             }
         }
@@ -355,7 +355,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
             Assert.AreEqual(
                 "overwritten", 
                 Configuration.Driver
-                .FindElement(By.TagName("input")).GetProperty("value")
+                .FindElement(By.TagName("input")).GetDomProperty("value")
             );
         }
 
@@ -409,7 +409,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneSpec
             Assert.AreEqual(
                 "overwritten", 
                 Configuration.Driver
-                .FindElement(By.TagName("input")).GetProperty("value")
+                .FindElement(By.TagName("input")).GetDomProperty("value")
             );
         }